+++ packages/terra-worklist-data-grid/src/proptypes/cellShape.js
@@ -2,9 +2,14 @@ import PropTypes from 'prop-types';
 
 const cellShape = PropTypes.shape({
   /**
-   * String indicating content to be displayed for the cell.
+   * String indicating content to be displayed for the cell.  Content will be ignored if isMasked property is true.
    */
   content: PropTypes.string,
+
+  /**
+   *  Boolean indicating if cell contents are masked.
+   */
+  isMasked: PropTypes.bool,
 });
 
 export default cellShape;