+++ packages/terra-worklist-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -1,6 +1,6 @@
 const selector = '#default-terra-worklist-data-grid';
 
-Terra.describeViewports('WorklistDataGrid', ['tiny', 'medium', 'large'], () => {
+Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
   describe('with keyboard navigation', () => {
     beforeEach(() => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/default-worklist-data-grid');
@@ -112,6 +112,12 @@ Terra.describeViewports('WorklistDataGrid', ['tiny', 'medium', 'large'], () => {
 
       Terra.validates.element('last-cell-in-last-row-selected', { selector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+    });
+
+    it('masked cells in worklist data grid', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-masked-cell');
+
+      Terra.validates.element('masked-cell', { selector });
     });
   });
 });