+++ packages/terra-framework-docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 
 * Added
   * Added documentation and tests for `terra-worklist-data-grid`.
+  * Added example and test for masked cell feature for `terra-worklist-data-grid`.
 
 ## 1.21.0 - (May 11, 2023)