+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added masked cell support.
+
 * Changed
   * Removed ability to set variable row heights.
   * Updated styling on cell focus.